Apply context, split component files
